backend/requirements/dev.txt
# Development requirements
pytest==7.0.0
pytest-flask==1.2.0
black==22.1.0
flake8==4.0.1